pyproject.toml
[project]
name = "spooklight"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "B.T. Franklin", email = "brandon.franklin@gmail.com"},
]
dependencies = [
    "click>=8.1.7",
    "openai>=1.42.0",
    "requests>=2.32.3",
    "colorama>=0.4.6",
    "promptdown>=0.7.5",
    "python-dotenv>=1.0.1",
    "reportlab>=4.2.2",
    "pillow>=10.4.0",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm.build]
excludes = ["tests/**"]

[tool.pdm.dev-dependencies]
dev = [
    "pytest>=8.3.2",
    "flake8>=7.1.1",
]
